requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==23.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.6
importlib-metadata==6.7.0
multidict==6.0.5
python-dotenv==0.21.1
typing_extensions==4.7.1
yarl==1.9.4
zipp==3.15.0

networkx~=3.2.1
pandas~=1.5.3
matplotlib~=3.6.3